+++ src/eval/evaluator_inc.hpp
@@ -38,7 +38,10 @@ class IncrementalEvaluator {
     INPUT_DEPENDENT_CELL_READ = 203,
     INPUT_DEPENDENT_SOURCE_USED = 204,
     NON_COMMUTATIVE_OPERATIONS = 205,
+    MEMORY_OP_SOURCE_INVALID = 210,
+    MEMORY_OP_WITH_INPUT_DEPENDENT_CELL_EXCEPT_COUNTER = 211,
   };
+
  public:
   explicit IncrementalEvaluator(Interpreter& interpreter);
 
@@ -88,7 +91,7 @@ class IncrementalEvaluator {
   void computeStatefulCells();
   void computeLoopCounterDependentCells();
   void initRuntimeData();
-  
+
   // Helper function to set error codes
   void setErrorCode(ErrorCode code, ErrorCode* error_code);