+++ src/eval/minimizer.cpp
@@ -86,9 +86,8 @@ bool Minimizer::minimize(Program& p, size_t num_terms) const {
           op.source.value != Number::ZERO) {
         int64_t base = getPowerOf(op.source.value);
         if (base != 0) {
-          std::unordered_set<int64_t> used_cells;
           int64_t largest_used = 0;
-          if (ProgramUtil::getUsedMemoryCells(p, used_cells, largest_used,
+          if (ProgramUtil::getUsedMemoryCells(p, nullptr, nullptr, largest_used,
                                               settings.max_memory)) {
             // try to replace gcd by a loop
             auto tmp = Operand(Operand::Type::DIRECT, largest_used + 1);